% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.R
\name{Load_LGR}
\alias{Load_LGR}
\title{Load_LGR}
\usage{
Load_LGR(file, Time_format = "mdy_hms")
}
\arguments{
\item{file}{set "file" as raw data file name (may include the directory path, as string) WITHOUT changing anything}

\item{Time_format}{depending on the output timestamp, you may change into "dmy_hms" or "ymd_hms", default: "mdy_hms"}
}
\value{
a dataframe with all the loaded data in the right format
}
\description{
Functions used to load raw data output from LGR Ultraportable Gas Analyzers to feed the function "FluxCal"
}
\examples{
Flux <- Load_LGR("C:/gga17Feb2017_f0000.txt")

}
