% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FluxCal.R
\name{FluxCal}
\alias{FluxCal}
\title{FluxCal}
\usage{
FluxCal(data, t, spt = 1, Time_keys = NA, Type_keys = "start", vol,
  Ta = NA, Area, output_d = "Flux_output.csv")
}
\arguments{
\item{data}{the data loaded by the function "Load_LGR" or "Load_other"}

\item{t}{window width for slope calculation, unit: minutes}

\item{spt}{set"spt", how many groups you would like to split your data into as you selet the peaks and valleys}

\item{Time_keys}{load a file with times (HH:MM:SS) that separate the flux calculation, with header as either "Start" or "End". if no file provided, a graph window will pop up to allow manually select the end points.}

\item{Type_keys}{select if "end" time or "start" time will be used to separate the flux; default "start"}

\item{vol}{volume of the chamber; unit: dm^3 or L}

\item{Ta}{air temperature; unit: degree C. A file name with a column "Ta" in the file. The length of Ta must be the same as number of measurements. Default: use the average ambient air temperature measured by analyzer}

\item{Area}{base area of the chamber; unit: m^2}

\item{output_d}{the output directory}
}
\value{
A dataframe with the calculated the fluxes
}
\description{
Calculate CO2 and CH4 gas flux from the data loaded by the function "Load_LGR" or "Load_other"
}
\examples{
    FluxCal(data = Flux, t = 3, vol = 1, Area = 1)

}
