% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.R
\name{Load_other}
\alias{Load_other}
\title{Load_other}
\usage{
Load_other(file, Time_format = "mdy_hms", Date = "1/31/2018", Time = NA,
  CO2, CH4 = NA, Ta = 25, skip = 0, sep = ",")
}
\arguments{
\item{file}{set "file" as raw data file name (may include the directory path, as string) WITHOUT changing anything}

\item{Time_format}{depending on the output timestamp, you may change into "dmy_hms" or "ymd_hms", default: "mdy_hms"}

\item{Date}{add the date of measurement in the format of "mm/dd/yyyy"}

\item{Time}{header name of the timestamps in the file (string)}

\item{CO2}{header name of the CO2 concentrations in the file (string)}

\item{CH4}{header name of the CH4 concentrations in the file (string). Default: no CH4 is measured}

\item{Ta}{header name of ambient air temperature in the file (string). If not measured, a constant value can be used. Default: 25 degree C}

\item{skip}{rows to be skipped at the beginning. Default: no rows will be skipped}

\item{sep}{delimiter to separate columns. Default: comma (",")}
}
\value{
a dataframe with all the loaded data in the right format
}
\description{
Functions used to load raw data output from other similar analyzers (e.g. LICOR) to feed the function "FluxCal"
}
\examples{
Flux <- Load_other("C:/Licor.txt",
                   Time_format = "hms",
                   Date = "5/11/2018",
                   Time = "time",
                   CO2 = "[co2]")

}
