% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.R
\name{Load_other}
\alias{Load_other}
\title{Load data from other analyzers}
\usage{
Load_other(file, Time_format = "mdy_hms", Date = NULL, Time, CO2,
  CH4 = NULL, Ta = 25, skip = 0, sep = ",")
}
\arguments{
\item{file}{A string contains the directory path and file name of the raw data.}

\item{Time_format}{A string indicates the format of timestamps. Default: "mdy_hms". It can be changed into "dmy_hms" or "ymd_hms", depending on the output timestamp from the analyzer. In case date and time are recorded separately, "hms" can be used here and add the date to the "Date" argument.}

\item{Date}{A string indicates date of the measurements in the format of "mm/dd/yyyy". This argument is used only when date is not included in the timestamps.}

\item{Time}{A string indicates column name of the timestamps in the data file.}

\item{CO2}{A string indicates column name of the CO2 concentrations in the data file.}

\item{CH4}{A string indicates column name of the CH4 concentrations in the data file. Default: no CH4 is measured.}

\item{Ta}{A string indicates column name of ambient air temperature in the data file. If not measured, a constant value can be used. Default: 25 degree C.}

\item{skip}{A number indicates rows to be skipped at the beginning of the data file. Default: 0, no rows will be skipped.}

\item{sep}{A string indicates delimiter to separate columns. Default: comma (",").}
}
\value{
a dataframe with all data in the right format for function "FluxCal()".
}
\description{
Function used to load raw data output from other similar analyzers (e.g. LICOR) to feed the "data" argument in function "FluxCal()".
}
\examples{
# get the directory of the example data
example_data <- system.file("extdata", "Flux_example_2_other.csv", package = "FluxCalR")
# load the data
Flux <- Load_other(example_data,
                   Time_format = "mdy_hms",
                   Time = "Date_time",
                   CO2 = "CO2_PPM",
                   Ta = "Tem_C")

}
